+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 1.1.0
- - Added baseline drift parameter to binding1to1 and binding2to1
+ - Added baseline drift parameter
+ - Added global offset parameter
+ - Added dissociation offset parameter
 
 # Version 1.0.0
  - First version.